# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/pull-requests:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: listSessionPullRequests
      tags: [Pull Requests]
      responses:
        "200":
          description: Normalized provider pull-request observations for the session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPullRequests"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/reviews:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: getSessionReviewState
      tags: [Reviews]
      responses:
        "200":
          description: AO review planning state and recorded runs for the session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionReviewState"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport, and
        `worker:notification` for durable agent notification intake.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PullRequestState:
      type: string
      enum: [draft, open, merged, closed]
    CIState:
      type: string
      enum: [unknown, pending, passing, failing]
    PullRequestCheckStatus:
      type: string
      enum: [unknown, queued, in_progress, passed, failed, skipped, cancelled]
    ReviewDecision:
      type: string
      enum: [none, approved, changes_requested, review_required]
    MergeabilityState:
      type: string
      enum: [unknown, mergeable, conflicting, blocked, unstable]
    PullRequestFailingCheck:
      type: object
      additionalProperties: false
      required: [name, status, conclusion]
      properties:
        name:
          type: string
        status:
          $ref: "#/components/schemas/PullRequestCheckStatus"
        conclusion:
          type: string
        url:
          type: string
    PullRequestCISummary:
      type: object
      additionalProperties: false
      required: [state, failingChecks]
      properties:
        state:
          $ref: "#/components/schemas/CIState"
        failingChecks:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestFailingCheck"
    PullRequestReviewCommentLink:
      type: object
      additionalProperties: false
      required: [autoInjectReview]
      properties:
        url:
          type: string
        file:
          type: string
        line:
          type: integer
          minimum: 0
        autoInjectReview:
          type: boolean
    PullRequestUnresolvedReviewer:
      type: object
      additionalProperties: false
      required: [reviewerId, count, links]
      properties:
        reviewerId:
          type: string
        count:
          type: integer
          minimum: 0
        links:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestReviewCommentLink"
        reviewUrl:
          type: string
        isBot:
          type: boolean
    PullRequestSubmittedReview:
      type: object
      additionalProperties: false
      required: [reviewerId, verdict, submittedAt, autoInjectReview]
      properties:
        reviewerId:
          type: string
        verdict:
          $ref: "#/components/schemas/ReviewDecision"
        body:
          type: string
        reviewUrl:
          type: string
        submittedAt:
          type: string
          format: date-time
        isBot:
          type: boolean
        autoInjectReview:
          type: boolean
    PullRequestReviewSummary:
      type: object
      additionalProperties: false
      required: [decision, hasUnresolvedHumanComments, unresolvedBy, reviews]
      properties:
        decision:
          $ref: "#/components/schemas/ReviewDecision"
        hasUnresolvedHumanComments:
          type: boolean
        unresolvedBy:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestUnresolvedReviewer"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestSubmittedReview"
    PullRequestConflictFile:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path:
          type: string
        url:
          type: string
    PullRequestMergeabilitySummary:
      type: object
      additionalProperties: false
      required: [state, reasons, pullRequestUrl, conflictFiles]
      properties:
        state:
          $ref: "#/components/schemas/MergeabilityState"
        reasons:
          type: array
          items:
            type: string
        pullRequestUrl:
          type: string
        conflictFiles:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestConflictFile"
    PullRequestSummary:
      type: object
      additionalProperties: false
      required:
        - url
        - number
        - title
        - state
        - provider
        - repository
        - author
        - sourceBranch
        - targetBranch
        - headSha
        - additions
        - deletions
        - changedFiles
        - ci
        - review
        - mergeability
        - updatedAt
        - observedAt
        - ciObservedAt
        - reviewObservedAt
      properties:
        url:
          type: string
        htmlUrl:
          type: string
        number:
          type: integer
          minimum: 1
        title:
          type: string
        state:
          $ref: "#/components/schemas/PullRequestState"
        provider:
          type: string
        repository:
          type: string
        author:
          type: string
        authorAvatarUrl:
          type: string
        sourceBranch:
          type: string
        targetBranch:
          type: string
        headSha:
          type: string
        additions:
          type: integer
          minimum: 0
        deletions:
          type: integer
          minimum: 0
        changedFiles:
          type: integer
          minimum: 0
        ci:
          $ref: "#/components/schemas/PullRequestCISummary"
        review:
          $ref: "#/components/schemas/PullRequestReviewSummary"
        mergeability:
          $ref: "#/components/schemas/PullRequestMergeabilitySummary"
        stateChangedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        observedAt:
          type: string
          format: date-time
        ciObservedAt:
          type: string
          format: date-time
        reviewObservedAt:
          type: string
          format: date-time
    SessionPullRequests:
      type: object
      additionalProperties: false
      required: [sessionId, pullRequests]
      properties:
        sessionId:
          type: string
        pullRequests:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestSummary"
    AOReviewRunStatus:
      type: string
      enum: [running, complete, delivered, failed, cancelled]
    AOReviewVerdict:
      type: string
      enum: ["", approved, changes_requested]
    AOReviewState:
      type: string
      enum: [needs_review, running, up_to_date, changes_requested, ineligible]
    AOReviewRun:
      type: object
      additionalProperties: false
      required:
        - id
        - reviewId
        - sessionId
        - batchId
        - harness
        - pullRequestUrl
        - targetSha
        - status
        - verdict
        - body
        - providerReviewId
        - createdAt
        - autoInjectReview
      properties:
        id:
          type: string
        reviewId:
          type: string
        sessionId:
          type: string
        batchId:
          type: string
        harness:
          type: string
        pullRequestUrl:
          type: string
        targetSha:
          type: string
        status:
          $ref: "#/components/schemas/AOReviewRunStatus"
        verdict:
          $ref: "#/components/schemas/AOReviewVerdict"
        body:
          type: string
        providerReviewId:
          type: string
        createdAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        autoInjectReview:
          type: boolean
    AOPullRequestReviewState:
      type: object
      additionalProperties: false
      required: [pullRequestUrl, pullRequestNumber, title, targetSha, status]
      properties:
        pullRequestUrl:
          type: string
        pullRequestNumber:
          type: integer
          minimum: 1
        title:
          type: string
        targetSha:
          type: string
        status:
          $ref: "#/components/schemas/AOReviewState"
        staleTargetSha:
          type: string
          description: Head SHA of the newest completed run that does not match targetSha.
        latestRun:
          $ref: "#/components/schemas/AOReviewRun"
        previousRun:
          $ref: "#/components/schemas/AOReviewRun"
    SessionReviewState:
      type: object
      additionalProperties: false
      required: [sessionId, reviews, runs]
      properties:
        sessionId:
          type: string
        reviewerHandleId:
          type: string
        reviewerHarness:
          type: string
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/AOPullRequestReviewState"
        runs:
          type: array
          items:
            $ref: "#/components/schemas/AOReviewRun"
...